+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.7] - 2021-03-11
+
+### Added
+
+  - Tests to `EngineBaseClient`.
+  - Tests to `SqliteClient`.
+
 ## [1.0.6] - 2021-03-01
 
 Add tests to `BigQueryClient`